+++ test/test.bash
@@ -8,7 +8,7 @@ dir=~
 cd $dir/ros2_ws
 colcon build
 source $dir/.bashrc
-timeout 1 ros2 launch mypkg prime_generator.launch.py > /tmp/mypkg.log
+timeout 1 ros2 launch mypkg prime_checker.launch.py > /tmp/mypkg.log
 
 cat /tmp/mypkg.log |
 grep 'prime'
